init8

+++ src/PayrollEngine.Domain/Entities/ActiveSSParams.cs
@@ -5,7 +5,7 @@ namespace PayrollEngine.Domain.Entities;
 public class ActiveSSParams
 {
     public Guid Id { get; set; }
-    public decimal Year { get; set; }
+    public int Year { get; set; }
     public decimal EmployeeSSRate { get; set; }
     public decimal EmployeeUIRate { get; set; }
     public decimal EmployerSSRate { get; set; }
