+++ src/PayrollEngine.Domain/Entities/CumulativeIncomeTaxBase.cs
@@ -5,6 +5,6 @@ namespace PayrollEngine.Domain.Entities;
 public class CumulativeIncomeTaxBase
 {
     public Guid Id { get; set; }
-    public decimal Month { get; set; }
+    public int Month { get; set; }
     public decimal CumulativeBase { get; set; }
 }
