+++ src/PayrollEngine.Domain/Entities/MinimumWage.cs
@@ -5,7 +5,7 @@ namespace PayrollEngine.Domain.Entities;
 public class MinimumWage
 {
     public Guid Id { get; set; }
-    public decimal Year { get; set; }
+    public int Year { get; set; }
     public decimal GrossAmount { get; set; }
     public decimal NetAmount { get; set; }
 }
