+++ src/PayrollEngine.Domain/Entities/SSCeiling.cs
@@ -5,6 +5,6 @@ namespace PayrollEngine.Domain.Entities;
 public class SSCeiling
 {
     public Guid Id { get; set; }
-    public decimal Year { get; set; }
+    public int Year { get; set; }
     public decimal Ceiling { get; set; }
 }
