+++ src/PayrollEngine.Infrastructure/Seed/ActiveSSParamsSeeder.cs
@@ -5,7 +5,7 @@ namespace PayrollEngine.Infrastructure.Seed;
 
 public class ActiveSSParamsSeeder
 {
-    public static void Seed(PayrollEngineDbContext context, decimal year, 
+    public static void Seed(PayrollEngineDbContext context, int year, 
                                                             decimal employeeSSRate, 
                                                             decimal employeeUIRate, 
                                                             decimal employerSSRate,
